created new class for dataproviders, created static methods for creating JSON requests
