+++ QA/PoroDevTest/src/test/java/api_test/negative_tests/DeleteUser.java
@@ -2,6 +2,7 @@ package api_test.negative_tests;
 
 import common.api_setup.ApiConfig;
 import common.api_setup.Endpoints;
+import common.api_setup.api_common.DataProviderBeUtil;
 import common.api_setup.api_common.UserDetailsGenerator;
 import io.restassured.specification.RequestSpecification;
 import org.testng.annotations.Test;
@@ -15,14 +16,11 @@ public class DeleteUser extends ApiConfig {
     public static RequestSpecification api_requestSpec;
 
     // Sending the
-    @Test(dataProvider = "invalidEmailList", dataProviderClass = UserDetailsGenerator.class)
+    @Test(dataProvider = "invalidEmailList", dataProviderClass = DataProviderBeUtil.class)
     public void deleteUserByInvalidEmail(String invalidEmailList) {
-        String jsonRequestWithEmail = "{\n" +
-                "  \"email\": \""+invalidEmailList+"\"\n" +
-                "}";
         given().relaxedHTTPSValidation()
                 .when()
-                .body(jsonRequestWithEmail)
+                .body(UserDetailsGenerator.createEmailJsonReq(invalidEmailList))
                 .delete(Endpoints.USER_DELETE)
                 .then()
                 .statusCode(anyOf(is(400), is(404), is(500)));
@@ -32,14 +30,11 @@ public class DeleteUser extends ApiConfig {
     // Sending delete request with valid but non-existing email
     @Test
     public void deleteUserByNonExistingEmail() {
-        String jsonRequestWithEmail = "{\n" +
-                "  \"email\": \"zivko@boing.rs\"\n" +
-                "}";
         given().relaxedHTTPSValidation()
                 .when()
-                .body(jsonRequestWithEmail)
+                .body(UserDetailsGenerator.createEmailJsonReq("zivko@boing.rs"))
                 .delete(Endpoints.USER_DELETE)
                 .then()
-                .statusCode(404);
+                .statusCode(anyOf(is(404), is(400)));
     }
 }
